Particle new texture

+++ sugiEngine/app/enemy/Enemy.h
@@ -20,7 +20,7 @@ private:
 private:
 	const float SPEED_MOVE = 0.15f;
 	const float SPEED_ANGLE = 5;
-	const uint32_t MAX_HP = 25;
+	const float MAX_HP = 25;
 	const float HEIGHT_COL = 2.2f;
 	const float SUB_TIME_ATTACK = 0.025f;
 	const float SPEED_DOWN = 5;
